Refactor algebra.h
