+++ includes/defines.h
@@ -6,7 +6,7 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/26 16:23:35 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/20 18:22:37 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/12/28 17:43:23 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,12 +26,5 @@
 
 # define BRIGHTNESS_MIN 0.0
 # define BRIGHTNESS_MAX 1.0
-
-/******************************************************************************\
- * COLORS
-\******************************************************************************/
-
-# define COLOR_MIN 0
-# define COLOR_MAX 255
 
 #endif
