+++ sources/mandatory/algebra/c3d.c
@@ -6,11 +6,11 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/18 22:55:39 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/26 15:03:49 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/12/28 17:42:53 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <minirt.h>
+#include <algebra.h>
 
 t_c3d	color_3d(double red, double green, double blue)
 {
