+++ sources/mandatory/algebra/magnitude.c
@@ -1,25 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   meta.c                                             :+:      :+:    :+:   */
+/*   magnitude.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/04/03 16:49:56 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/28 16:36:48 by lpaulo-m         ###   ########.fr       */
+/*   Created: 2022/12/28 17:42:18 by lpaulo-m          #+#    #+#             */
+/*   Updated: 2022/12/28 17:42:53 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <minirt.h>
-
-t_t3d	neg(t_t3d tuple)
-{
-	tuple.x *= -1;
-	tuple.y *= -1;
-	tuple.z *= -1;
-	tuple.type *= -1;
-	return (tuple);
-}
+#include <algebra.h>
 
 double	length_squared(t_v3d vector)
 {
@@ -46,25 +37,4 @@ double	length(t_v3d vector)
 double	magnitude(t_v3d vector)
 {
 	return (length(vector));
-}
-
-t_v3d	unit(t_v3d vector)
-{
-	double	magnitude;
-
-	magnitude = length(vector);
-	return (scalar_div(magnitude, vector));
-}
-
-t_v3d	normalize(t_v3d vector)
-{
-	return (unit(vector));
-}
-
-t_v3d	vsqrt(t_v3d vector)
-{
-	vector.x = sqrt(vector.x);
-	vector.y = sqrt(vector.y);
-	vector.z = sqrt(vector.z);
-	return (vector);
 }
