+++ sources/mandatory/algebra/meta_vectors.c
@@ -1,32 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   geometry.c                                         :+:      :+:    :+:   */
+/*   meta_vectors.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2021/04/03 16:34:17 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/28 17:47:33 by lpaulo-m         ###   ########.fr       */
+/*   Created: 2021/04/03 16:49:56 by lpaulo-m          #+#    #+#             */
+/*   Updated: 2022/12/28 17:48:42 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <algebra.h>
 
-static double	abs_d(double number)
+t_v3d	unit(t_v3d vector)
 {
-	if (number < 0)
-		return (number * -1);
-	return (number);
+	double	magnitude;
+
+	magnitude = length(vector);
+	return (scalar_div(magnitude, vector));
 }
 
-bool	near_zero(t_v3d vector)
+t_v3d	normalize(t_v3d vector)
 {
-	bool	cuts_off_x;
-	bool	cuts_off_y;
-	bool	cuts_off_z;
+	return (unit(vector));
+}
 
-	cuts_off_x = abs_d(vector.x) < NEAR_ZERO_TOLERANCE;
-	cuts_off_y = abs_d(vector.y) < NEAR_ZERO_TOLERANCE;
-	cuts_off_z = abs_d(vector.z) < NEAR_ZERO_TOLERANCE;
-	return (cuts_off_x && cuts_off_y && cuts_off_z);
+t_v3d	vsqrt(t_v3d vector)
+{
+	vector.x = sqrt(vector.x);
+	vector.y = sqrt(vector.y);
+	vector.z = sqrt(vector.z);
+	return (vector);
 }
