+++ sources/mandatory/algebra/scalar.c
@@ -6,11 +6,11 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/03 16:34:17 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/28 16:37:03 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/12/28 17:49:01 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <minirt.h>
+#include <algebra.h>
 
 t_t3d	scalar_add(double number, t_t3d tuple)
 {
@@ -46,18 +46,4 @@ t_t3d	scalar_div(double number, t_t3d tuple)
 	tuple.z /= number;
 	tuple.type /= number;
 	return (tuple);
-}
-
-double	dot(t_v3d first, t_v3d second)
-{
-	double	x;
-	double	y;
-	double	z;
-	double	type;
-
-	x = first.x * second.x;
-	y = first.y * second.y;
-	z = first.z * second.z;
-	type = first.type * second.type;
-	return (x + y + z + type);
 }
