+++ sources/mandatory/algebra/tuple.c
@@ -6,11 +6,11 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/26 15:11:13 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/26 19:19:37 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/12/28 17:48:50 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <minirt.h>
+#include <algebra.h>
 
 t_t3d	tuple(double x, double y, double z, double type)
 {
@@ -30,4 +30,13 @@ t_t3d	null_tuple(void)
 void	inspect_tuple(t_t3d tuple)
 {
 	printf("TUPLE: %f %f %f, type %f\n", tuple.x, tuple.y, tuple.z, tuple.type);
+}
+
+t_t3d	neg(t_t3d tuple)
+{
+	tuple.x *= -1;
+	tuple.y *= -1;
+	tuple.z *= -1;
+	tuple.type *= -1;
+	return (tuple);
 }
